Move conversion from VersionedIdentifier to CqlVersionedLibraryidentifier to extensions type
